backend/requirements.txt
Django>=4.1.3,<4.2
djangorestframework>=3.14.0,<3.15
autopep8>=2.0.0,<2.1
djoser>=2.1.0,<2.2
djangorestframework_simplejwt>=4.8.0,<4.9
drf-yasg>=1.21.4,<1.22
django-cors-headers>=3.13,<3.14
coverage>=7.1.0
black>=23.1.0